pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "judgebench"
version = "0.1.0"
description = "LLM judge quality evaluator with bias detection and calibration dashboard"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "httpx>=0.27",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
judgebench = "judgebench.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/judgebench"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
